+++ complex.h
@@ -4,11 +4,11 @@
 typedef struct complex *complex;
 
 /*initial complex number*/
-complex init_comp(float real, float img);
+complex init_comp(double real, double img);
 
 
  /*get a complex number and change its rael and img to be the new real and img*/
- void read_comp(complex complex1,float new_real,float new_img);
+ void read_comp(complex complex1,double new_real,double new_img);
 
 /*get a complex number to print and print it*/
 void print_comp(complex to_print);
@@ -20,10 +20,10 @@ void add_comp(complex dest, complex complex1, complex complex2);
 void sub_comp(complex dest, complex complex1, complex complex2);
 
 /*get complex number and real number and return their multiplication result*/
-void mult_comp_real(complex dest, complex complex1, float real);
+void mult_comp_real(complex dest, complex complex1, double real);
 
 /*get complex number and imaginary number and assign their multiplication result to dest*/
-void mult_comp_img(complex dest, complex complex1, float img);
+void mult_comp_img(complex dest, complex complex1, double img);
 
 /*get tow complex numbers and assign their multiplication result to dest*/
 void mult_comp_comp(complex dest, complex complex1, complex complex2);
